Really tiny update for RefactorItem class
Moved to a separate file, and changed Description property value.
